WGML change justification default to OFF      fix signed / unsigned calculation for .sk -1
[git-p4: depot-paths = "//depot/openwatcom/": change = 37456]

+++ bld/wgml/c/gssk.c
@@ -46,7 +46,7 @@ int32_t calc_skip_value( void )
 {
     int32_t     skip_value;
 
-    skip_value = (g_skip * bin_device->vertical_base_units ) / LPI;
+    skip_value = (g_skip * (int32_t)bin_device->vertical_base_units ) / LPI;
     g_skip = 0;
     return( skip_value );
 }
